requirements.txt
Package             Version
------------------- --------
Django              >=4.1.0, <4.2
djangorestframework >=3.14.0, <3.15
django-filter       >=22.1, <22.2
django-crispy-forms >=1.14, <1.14.1

